Checkpoint after Furbolg battle

+++ core/src/com/etheller/warsmash/viewer5/handlers/w3x/simulation/CDestructableType.java
@@ -8,14 +8,14 @@ import com.etheller.warsmash.viewer5.handlers.w3x.simulation.combat.CTargetType;
 public class CDestructableType {
 
 	private final String name;
-	private final int life;
+	private final float life;
 	private final EnumSet<CTargetType> targetedAs;
 	private final String armorType;
 	private final int buildTime;
 	private final BufferedImage pathingPixelMap;
 	private final BufferedImage pathingDeathPixelMap;
 
-	public CDestructableType(final String name, final int life, final EnumSet<CTargetType> targetedAs,
+	public CDestructableType(final String name, final float life, final EnumSet<CTargetType> targetedAs,
 			final String armorType, final int buildTime, final BufferedImage pathingPixelMap,
 			final BufferedImage pathingDeathPixelMap) {
 		this.name = name;
@@ -31,7 +31,7 @@ public class CDestructableType {
 		return this.name;
 	}
 
-	public int getLife() {
+	public float getLife() {
 		return this.life;
 	}
